python/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
optimizely-sdk==5.0.0
pycparser==2.22
pyOpenSSL==24.1.0
pyrsistent==0.20.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
urllib3==2.2.1
Werkzeug==3.0.3
